Split some Obj-C-related code from cocoa_webkit.hh
